# pipelineService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/destinations/": {
            "get": {
                "description": "Return all the destinations supported by cdpaas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "destination"
                ],
                "summary": "Get All Supported Destinations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SupportedDestinationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Configures and creates a destination connector on Airbyte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "destination"
                ],
                "summary": "Create Destination on Airbyte and store related information in local database.",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateDestinationConnectorRequestAPI"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreateDestinationConnectorResponseAPI"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/destinations/configured/": {
            "get": {
                "description": "Return all the destinations that are configured on cdpaas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "destination"
                ],
                "summary": "Get All Configured Destinations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConfiguredDestinationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/destinations/specification/": {
            "get": {
                "description": "Return the specification of a given destination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "destination"
                ],
                "summary": "Get Destination Specification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination Name",
                        "name": "destination",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DestinationSpecification"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/destinations/{id}/summary/": {
            "get": {
                "description": "Returns the destination's summary",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "destination"
                ],
                "summary": "Get Destination Summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DestinationSummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ConfiguredDestination": {
            "type": "object",
            "required": [
                "destinationId",
                "destinationName",
                "destinationType"
            ],
            "properties": {
                "airbyteDestinationDefinitionId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "airbyteDestinationId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "destinationId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "destinationName": {
                    "type": "string",
                    "example": "example_destination"
                },
                "destinationType": {
                    "type": "string",
                    "example": "example_destination"
                },
                "host": {
                    "type": "string"
                }
            }
        },
        "models.ConfiguredDestinationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConfiguredDestination"
                    }
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "models.CreateDestinationConnectorRequestAPI": {
            "type": "object",
            "required": [
                "connectionConfiguration",
                "destinationDefinitionId",
                "destinationName",
                "name"
            ],
            "properties": {
                "connectionConfiguration": {
                    "type": "string"
                },
                "destinationDefinitionId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "destinationName": {
                    "type": "string",
                    "example": "Postgres"
                },
                "name": {
                    "type": "string",
                    "example": "example_destination_name"
                }
            }
        },
        "models.CreateDestinationConnectorResponseAPI": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.CreateDestinationConnectorResponseData"
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.CreateDestinationConnectorResponseData": {
            "type": "object",
            "properties": {
                "destinationId": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "destinationName": {
                    "type": "string",
                    "example": "Postgres"
                },
                "name": {
                    "type": "string",
                    "example": "example_destination"
                }
            }
        },
...
        "models.DestinationSpecification": {
            "type": "object",
            "properties": {
                "advancedAuth": {
                    "type": "object"
                },
                "authSpecification": {
                    "type": "object"
                },
                "connectionSpecification": {
                    "type": "object"
                },
                "destinationDefinitionId": {
                    "type": "string"
                },
                "documentationUrl": {
                    "type": "string"
                },
                "jobInfo": {
                    "type": "object"
                },
                "supportedDestinationSyncModes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "supportsDbt": {
                    "type": "boolean"
                },
                "supportsNormalization": {
                    "type": "boolean"
                }
            }
        },
        "models.DestinationSummaryResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "destinationName",
                "owner"
            ],
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "destinationName": {
                    "type": "string",
                    "example": "example_destination"
                },
                "owner": {
                    "type": "object"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.SupportedDestinationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SupportedDestinations"
                    }
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.SupportedDestinations": {
            "type": "object",
            "required": [
                "id",
                "name",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "name": {
                    "type": "string",
                    "example": "destination name"
                },
                "type": {
                    "type": "string",
                    "example": "destination type"
                }
            }
        },
...
    }
}